refactor: 아키텍처에 맞춰 패키지 구조 변경
