+++ src/main/java/hhplus/lecture/domain/lectureApply/LectureApplyEntity.java
@@ -1,6 +1,6 @@
 package hhplus.lecture.domain.lectureApply;
 
-import hhplus.lecture.domain.AuditingFields;
+import hhplus.lecture.domain.util.AuditingFields;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
