+++ src/main/java/hhplus/lecture/domain/util/AuditingFields.java
@@ -1,4 +1,4 @@
-package hhplus.lecture.domain;
+package hhplus.lecture.domain.util;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
@@ -22,12 +22,12 @@ public class AuditingFields {
 
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     @CreatedDate
-    @Column(name = "REG_DT", nullable = false, updatable = false)
-    private LocalDateTime regDt; // 등록일시
+    @Column(name = "created_at", nullable = false, updatable = false)
+    private LocalDateTime createdAt; // 등록일시
 
     @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
     @LastModifiedDate
-    @Column(name = "MDFCN_DT", nullable = false)
-    private LocalDateTime mdfcnDt; //수정일시
+    @Column(name = "modified_at", nullable = false)
+    private LocalDateTime modifiedAt; //수정일시
 
 }
