+++ src/main/java/hhplus/lecture/domain/util/Constants.java
@@ -1,4 +1,4 @@
-package hhplus.lecture.config;
+package hhplus.lecture.domain.util;
 
 public final class Constants {
     public static final Integer MAX_APPLICANTS = 30;
