+++ src/main/resources/application.yaml
@@ -37,9 +37,9 @@ spring:
       on-profile: test
   datasource:
     driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:testdb
-    username: test
-    password: test1234
+    url: jdbc:h2:mem:testdb;MODE=MySQL;DATABASE_TO_UPPER=false
+    username: sa
+    password:
     jpa:
       hibernate:
         ddl-auto: create-drop
